refactor: move s050 migration into separate file
