+++ attest/Cargo.toml
@@ -40,7 +40,7 @@ tracing.workspace = true
 url = "2.2"
 zbus.workspace = true
 zenorb.workspace = true
-zbus_systemd = { workspace = true, features = ["systemd1"] }
+zbus_systemd = { workspace = true, features = ["systemd1"], optional = true }
 tempfile = "3.3"
 
 [build-dependencies]
@@ -57,4 +57,4 @@ systemd-units = [{ unit-name = "worldcoin-attest" }]
 
 [features]
 default = ["se050_key_migration"]
-se050_key_migration = []
+se050_key_migration = ["dep:zbus_systemd"]
